pyproject.toml
[project]
name = "proyectoo1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-pubsub>=2.35.0",
]
